App routing hox fix bug
