+++ web/src/navigation/profile-routes.ts
@@ -1,6 +1,6 @@
 import Profile from "@/pages/profile";
 import { UserCogIcon } from "lucide-react";
-import { createLeafRoute } from "./app-routes";
+import { createLeafRoute } from "./app-route-factory";
 
 export const ProfileRoute = createLeafRoute({
   path: "/profile",
