+++ web/src/navigation/system-routes.ts
@@ -2,7 +2,7 @@ import Login from "@/pages/Login";
 import NotFound from "@/pages/NotFound";
 import Unauthorized from "@/pages/Unauthorized";
 import { LogInIcon, ShieldBanIcon, TriangleAlertIcon } from "lucide-react";
-import { createLeafRoute } from "./app-routes";
+import { createLeafRoute } from "./app-route-factory";
 import { MedicalTestsRoute } from "./medical-tests-routes";
 
 export const LoginRoute = createLeafRoute({
